feat(.zshrc): add alias 'tagn' to display git tags with annotations
The 'tagn' alias is added to the .zshrc file to provide a convenient way to display git tags along with their annotations. This alias executes the 'git tag -n' command, which shows the tags and their corresponding annotations. This addition enhances the productivity of the user by providing a quick way to view the annotations associated with git tags.